pyproject.toml
[project]
name = "eoir"
version = "0.1.0"
description = "EOIR FOIA data processing tools"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.1.8",
    "psycopg==3.2.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "structlog>=24.1.0",
    "zipfile-deflate64>=0.2.0",
    "ipython"
]

[project.scripts]
eoir = "eoir.main:cli"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["eoir"]
